Allow concurrency to go to 1
It looks like this was set to 2 because it could cause deadlocks at
lower numbers. I think the problem is that the semaphore was counting up
while walking the directory tree (before starting the test goroutine).

Because walking the directory tree would block, testFiles() would never
drain the test output channel, and the tests would deadlock.

This change moves the semaphore into the test goroutine, so
walkDirectories is sure to return.