Allow overalls to work with projects based on Go modules
Previously overalls wouldn't work with a project that used Go modules.
It uses go/build.Import to locate the package identified by the project
path in the input, but since an empty source path was passed we never
entered the code path that dealt with modules so we ended up searching
GOPATH directories only (and with modules, of course, GOPATH isn't set).

Additionally, we now use the resolved package's actual source path
instead of using the source root (typically $GOPATH/src unless modules
are used) and making assumptions about the source path.